% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_reg.R
\name{make_reg}
\alias{make_reg}
\title{Utility function to make classical calibration equations}
\usage{
make_reg()
}
\value{
A four row by three column matrix, where the rows are for femur,
  tibia, humerus, and radius, and the columns are for slope, intercept, and
  SEE.
}
\description{
This function takes the vector of means \link{mu} and the
  variance- covariance matrix \link{V.2015} from Konigsberg and Jantz (2018,
  Table 9.2, first sample) and calculates the intercepts, slopes (equation
  2), and standard error of estimates (SEEs) (square root of IMSE from Table
  3).
}
\examples{
\dontrun{
# In package, output is stored in K_reg
make_reg()
}
}
\references{
Konigsberg, L. W., & Meadows Jantz, L. (2018). Multivariate
regression methods for the analysis of stature. In K. Latham, E. Bartelink,
 & M. Finnegan (Eds.), \emph{New Perspectives in Forensic Human Skeletal
 Identification} (pp. 87-104). Cambridge, MA: Academic Press.

Konigsberg, L. W., Hens, S. M., Jantz, L. M., & Jungers, W. L. (1998).
Stature estimation and calibration: Bayesian and maximum likelihood
perspectives in physical anthropology. \emph{Yearbook of Physical
Anthropology}, 41, 65-92.
}
