% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mu}
\alias{mu}
\title{Vector of means}
\format{
A 1 by 5 row vector
}
\usage{
data(mu)
}
\description{
This data is a row vector of means for stature, maximum femur,
tibia, humerus, and radius length (all in mm.) from Konigsberg and Jantz (2018).
}
\references{
Konigsberg, L. W., & Meadows Jantz, L. (2018). Multivariate regression
 methods for the analysis of stature. In K. Latham, E. Bartelink,
 & M. Finnegan (Eds.), \emph{New Perspectives in Forensic Human Skeletal
 Identification} (pp. 87-104). Cambridge, MA: Academic Press.
}
\keyword{datasets}
