% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{S2_reg}
\alias{S2_reg}
\title{Sjøvold (1990) regression data}
\format{
A data frame with 4 rows and 3 columns where
the rows are for maximum femur, tibia, humerus, and radius lengths
and the columns are the regression coefficient, the regression
constant, and the standard error of estimate. The bone measurements
should be in mm. and the returned stature estimates are in mm.
}
\usage{
data(S2_reg)
}
\description{
Contains the bivariate regression (RMA) data for stature on long bones
from Sjøvold (1990) Table 2.  Note: This is for "all ethnic groups independent
of sex
}
\references{
Sjøvold, T. (1990). Estimation of stature from long bones utilizing
the line of organic correlation. \emph{Human Evolution}, 5, 431-447.
}
\keyword{datasets}
